Add minor changes in devotion routes and admindhasboard
